pyproject.toml
[project]
name = "generative_saleman"
version = "0.1.0"
description = "This is our class project for the course of NLP SYSTEMS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.24",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.34",
    "mcp-openai>=0.2.0",
    "mcp[cli]>=1.5.0",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.9",
    "pyyaml>=6.0.2",
    "pyzbar>=0.1.9",
    "streamlit>=1.44.0",
    "supabase>=2.15.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[tool.pytest]
addopts = ["--cov=generative_saleman", "--cov-report=term-missing"]
testpaths = ["tests"]

[tool.pytest.ini_options]
python_files = "test_*.py"